requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-marshmallow
flask-login
pytest